move getFullSwapERC20 to swap-erc20 to fix circular dep
