+++ source/staking/package.json
@@ -31,7 +31,7 @@
     "access": "public"
   },
   "devDependencies": {
-    "@airswap/utils": "^4.1.1",
+    "@airswap/utils": "^4.1.2",
     "prompt-confirm": "^2.0.4"
   }
 }
