+++ source/swap-erc20/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@airswap/swap-erc20",
-  "version": "4.1.0",
+  "version": "4.1.1",
   "description": "AirSwap: Atomic ERC20 Token Swap",
   "license": "MIT",
   "repository": {
@@ -31,7 +31,7 @@
     "@airswap/constants": "^4.1.0",
     "@airswap/staking": "4.0.3",
     "@airswap/types": "^4.1.1",
-    "@airswap/utils": "^4.1.1",
+    "@airswap/utils": "^4.1.2",
     "prompt-confirm": "^2.0.4"
   },
   "publishConfig": {
