+++ source/swap/package.json
@@ -34,7 +34,7 @@
   "devDependencies": {
     "@airswap/constants": "^4.1.0",
     "@airswap/types": "^4.1.1",
-    "@airswap/utils": "^4.1.1",
+    "@airswap/utils": "^4.1.2",
     "@nomicfoundation/hardhat-network-helpers": "^1.0.7"
   },
   "publishConfig": {
