+++ tools/libraries/package.json
@@ -34,9 +34,9 @@
     "@airswap/pool": "4.1.0",
     "@airswap/staking": "4.0.3",
     "@airswap/swap": "4.0.4",
-    "@airswap/swap-erc20": "4.1.0",
+    "@airswap/swap-erc20": "4.1.1",
     "@airswap/types": "^4.1.1",
-    "@airswap/utils": "^4.1.1",
+    "@airswap/utils": "^4.1.2",
     "@airswap/wrapper": "4.1.0",
     "browser-or-node": "^2.1.1",
     "ethers": "^5.6.9",
