+++ tools/utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@airswap/utils",
-  "version": "4.1.1",
+  "version": "4.1.2",
   "description": "AirSwap: Utilities for Developers",
   "repository": {
     "type": "git",
@@ -22,7 +22,6 @@
   },
   "dependencies": {
     "@airswap/constants": "^4.1.0",
-    "@airswap/swap-erc20": "4.1.0",
     "@airswap/types": "^4.1.1",
     "@metamask/eth-sig-util": "^5.0.2",
     "bignumber.js": "^9.0.1",
